refactor: move schemas.ts
